feat(routing): change RpcOperation trait
To handle routes, we need distinct pieces of data, from distinct
sources. Change trait signature not requiring mutable borrow and
use RoutingDb as ObjectStore in most cases since that allows us
to access any piece of state needed. The downside is that we need
to take the locks within the handlers. This limitation will
disappear once we remove the locking.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>