feat(routing): let the CPI own the routing db
Let the CPI own the routing database since we need it to own
the fibtable writer object. This change is very significant as
it will allow us to remove most of the locking, simplifying the
code and making it more efficient. Such a refactor will come in
subsequent commits, once the fib is usable.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>