ci: Disable stable toolchain for dev workflow
We don't use the "stable" toolchain these days, given that we require
Rust features that are in 1.86 (the current beta). Running the stable
workflow in CI just burns resources and generates some noise in the
discussion thread for the Pull Requests. Disable it.

This commit should be revert right before we revert commit 643de7afc7fc
("fix(justfile): Use beta release of rust, not stable").

Note: 1.86 is due to become stable on 3rd April
(https://releases.rs/docs/1.86.0/).

Signed-off-by: Quentin Monnet <qmo@qmon.net>