chore: Add missing license info in Cargo.toml files
To avoid ambiguity in terms of licensing, explicitly provide a license
parameters in the Cargo.toml of all packages in the repository. This
should also help FOSSA, which currently struggles to determine the
license for these crates.

The license is set to Apache-2.0, as for the packages that already
stated a license.

Also mark the dataplane-id crate as non-publishable, for consistency
with the other packages.

Signed-off-by: Quentin Monnet <qmo@qmon.net>