docs: Add just target to serve generated rustdoc locally
Make it easy for users to find and browse the rustdoc pages that are
generated locally with "just cargo doc".

Note: It should be possible to run the pages without launching a server,
but I run my browser as a Flatpak app that doesn't have access to the
part of my filesystem where these docs are located. Launching a local
webserver with Python is the easiest workaround and should be available
to most users.

Signed-off-by: Quentin Monnet <qmo@qmon.net>