feat(routing): define HasMac/HasIfMapping traits
The HasMac trait defines a getter for the mac of an interface.
The HasIfMapping trait defines a mapping() method that provides
an IfMapping object that should be unique per interface. The
intent is to use the IfMapping to map an incoming packet to the
corresponding kernel interface. Without subinterfaces, an IfMapping
could just be a Mac address. However, to support for 8021.1q, we
also add a vlan id. Other fields may be needed depending on the
encapsulations used.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>