ci(sterile): Disable "cargo doc" with release profile
We currently have an issue with the release profile: doctests do not
build correctly. This is apparently related to LTO, as the issue
disappears if we turn it off.

It seems OK to skip "cargo doc" with the release profile, as long as we
make sure it passes with the dev profile. Let's remove that step from
the workflow.

Signed-off-by: Quentin Monnet <qmo@qmon.net>