build: Set RUSTDOCFLAGS to pass linker flags for release profile
See #362 on GitHub for the full context. In short, we're not passing the
linker flags with RUSTFLAGS when trying to build the doctests, and this
is an issue when LTO is turned on (as is the case with the "release"
profile), because we fall back to cc/ld as the default linker, which
results in a failure to build the doctests, ld being unable to process
the LLVM IR it gets due to LTO being enabled.

Make sure we pass the relevant flags via RUSTDOCFLAGS.

Signed-off-by: Quentin Monnet <qmo@qmon.net>