build: Revert to -v3 instruction set on x86-64
Some developers experience issues when trying to compile with the flag
"-march=x86-64-v4" on their developement setup, because their CPU is not
recent enough. To ensure that we can work in good conditions, change it
to v3. Note that this is what we do in the compile-env already, since
commit githedgehog/dpdk-sys@723c4656324c.

Signed-off-by: Quentin Monnet <qmo@qmon.net>