ci: Make sure to pass env variables to sterile environment
On my system, it appears that CARGO_TARGET_DIR is not passed to the
sterile container's environment (I suspect DOCKER_HOST is set inside of
the container and not passed through the command-line, either). I
believe that the sudoers file on my machine only allows a few specific
environment variables to be shared with "sudo -E", and CARGO_TARGET_DIR
does not belong to the list.

Make sure to pass the environment variables as expected, by specifying
their values (the variables are expanded before the command is run in
the sudo environment).

Signed-off-by: Quentin Monnet <qmo@qmon.net>