ci: Pass TMPDIR to sterile environment
We have an issue with doctests in CI: "cargo test --doc" fails.

Investigation revealed that:

- The doctests fail to compile, making the process return with 101
- 101 is the error code for rustc or rustdoc when they panic
- The doctests compile and pass just fine outside of the sterile
  container environment, including when using the same binaries
- Interestingly, "strace -f -e execve" on the rustdoc command showed
  that rustdoc attempts to generate a binary it generated under /tmp/,
  and that this execution fails due to missing permissions.

In other words, the issue occurs in the sterile environment because
rustdoc attempts to execute a binary located under /tmp/, that we mount
with the "noexec" flag in the container.

One workaround is to turn off "noexec", although this is not ideal.
Instead, we follow the suggestion from the relevant rust Issue: we pass
TMPDIR as an environment variable, to tell rustdoc to use a directory
where it can execute binaries rather than /tmp/.

Successfully tested, after re-enabling the doctests locally, with:

    just debug_justfile=true rust="beta" profile=dev \
        target=x86_64-unknown-linux-gnu sterile cargo test --doc

Issue: #362
Link: https://github.com/rust-lang/rust/issues/40065
Signed-off-by: Quentin Monnet <qmo@qmon.net>