fix(dataplane): fix kernel driver packet reception
Mio used to allow specifying in a PollOpt if a descriptor regis-
tration for polling would be edge or level triggered. The version
we use doesn't seem to allow for this and probably defaults to
edge. Therefore, when notified about data for reading, let the
kernel driver read data until an error occurs.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>