feat(dataplane): let PacketDumper have a name
A pipeline may be built to contain more than a packet dumper; e.g.
one right after some ingress stage and another one right before
egress. In order to know at which stage a packet was dumped, let
PacketDumper objects get a name that is displayed with the packet
dump.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>