feat(mgmt): Load GatewayConfig from YAML
This is a temporary commit meant to provide YAML loading for the
GatewayConfig object in the dataplane, as a temporary replacement for
the object we should get as a result from an update gRPC message sent by
the Agent.

The code defines the different objects for the gRPC requests and
messages, and provides a function to load a (root) GatewayConfig object
from a YAML file - currently with a hardcoded path.

File mgmt/src/grpc/config.rs was generated from the following program,
to process the dataplane.proto file from the gateway-proto repository:

    use tonic_build;

    fn main() {
        let proto = "proto/dataplane.proto";

        let res = tonic_build::configure()
            .type_attribute(".", "#[derive(::serde::Deserialize, ::serde::Serialize)]")
            .compile_protos(&[proto], &["proto"]);

        match res {
            Ok(_) => println!("Protobuf compiled successfully!"),
            Err(e) => {
                eprintln!("Failed to compile Protobuf: {}", e);
                std::process::exit(1);
            }
        }
    }

This small program uses tonic-build as a dependency:

    [dependencies]
    tonic-build = "0.13.0"

We can run this program with, for example:

    mkdir -p output
    PROTOC=/usr/bin/protoc OUT_DIR=output cargo run

This generates file output/config.rs, that we copied here.

Signed-off-by: Quentin Monnet <qmo@qmon.net>