ci: Don't error out of "bump" workflow when log file cannot be removed
At the end of the upgrade step in the "bump" GitHub workflow, we attempt
to remove the generated log files to avoid them being committed and
ending up in the generated Pull Request. However, one of these files may
occasionally be absent, causing the "rm" command to fail.

The issue was observed for example at
https://github.com/githedgehog/dataplane/actions/runs/14436825830/job/40479270201#step:8:110

Just ignore nonexistent files when cleaning up: "rm -f".

Signed-off-by: Quentin Monnet <qmo@qmon.net>