ci: In "bump" workflow, remove log files when exiting early
On early exit from the upgrade step in the "bump" GitHub workflow, we
would not clean up the generated log files. One consequence of it is
that the files may end up in the Pull Request generated at the next
workflow step.

This issue was observed, for example, in:
https://github.com/githedgehog/dataplane/pull/398/files

Let's make sure we clean up these files to avoid a pseudo-empty PR to be
generated.

Signed-off-by: Quentin Monnet <qmo@qmon.net>