fix(net/interface): Use serde traits only when "serde" feature is on
The serde dependency crate is optional for the "net" package, and we
cannot always rely on it. Otherwise, the build fails when we do not
enable the "serde" package feature. Fix the issue by using conditional
attributes, based on the state of the "serde" feature.

For more context, please refer to the previous commit, in which we
addressed a similar issue for net/ethtype.

Fixes: 55d2b2fc46b5 (feat: convenience methods for ethtype)
Fixes: 06678572fe32 (feat: intro interface types)
Reported-by: Fredi Raspall <fredi@githedgehog.com>
Signed-off-by: Quentin Monnet <qmo@qmon.net>