feat: add back nextest config
nextest is a more modern test runner / reporting tool than the one which ships with cargo.
nextest is significantly faster than cargo test and can report as junit et al.

To run nextest tests you can call

```
just cargo nextest run
```

It was used before, but I removed it in the name of making the justfile less complex.

Since then the justfile has been refactored to the point where nextest is very low cost to support, so we may as well add it back.  I like the faster tests.

Note: we can't currently support nextest in the sterile build env as I have yet to add it to the compile-env.  I may do this later if folks want it.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>