feat(mgmt): Add frr_reload.py wrapper
Add a wrapper to run the frr_reload.py Python script, necessary to
test then reload FRR's configuration. When running the script, we
collect the standard output and error, and the return code. Any data
printed to stderr, or non-zero return code, means that the script hasn't
executed as expected, and that the wrapper returns an error.

Signed-off-by: Quentin Monnet <qmo@qmon.net>