feat(mgmt): add faked frr-agent for testing
The introduction of keepalives and the fact that frr-agent is
probed with one keepalive when creating the frrmi require a more
elaborated logic to fake an frr-agent. Add a test util called
'fake_frr_agent' that spawns a tokio task to emulate an frr-agent.
The task simply binds a socket to the specified path and replies
with an "Ok" to any request.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>