test(test-runner): Per-package opt-in for test runner
The test runner that makes unit tests run inside of a Docker container
tend to make it harder to run and debug tests. Only use the container
for packages that do require specific configuration or elevated
privileges, not for the other packages.

As part of the change, make sure to undo the path change that was
necessary for the test runner in net/src/packet/hash.rs.

[ Quentin: Picked up Daniel's snippet from Slack and committed it. ]

Authored-by: Daniel Noland <daniel@githedgehog.com>
Signed-off-by: Quentin Monnet <qmo@qmon.net>