feat(mgmt): build blank config
We need to be able to apply a 'blank' config in order to flush the
system (e.g. to roll back an unsuccesful configuration if no prior
config was there).

This patch adds a builder for a blank config. Such a config has
a reserved generation id of 0 and it cannot be deleted. Since the
config is stored in the config database, where configs are keyed
by generation id, this means that attempting to add a config with
generation id 0 will fail.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>