feat(vpcpeering): Introduce validation for VpcExpose objects
Implement the skeleton function to validate a VpcExpose object. This
validation function is automatically called when adding a VpcExpose to a
VpcManifest object. The function performs the following steps:

 1. Make sure that all prefixes and exclusion prefixes for this
    VpcExpose are of the same IP version.
 2. Make sure that all prefixes (or exclusion prefixes) in each list
    (ips/nots/as_range/not_as) don't overlap with other prefixes (or
    exclusion prefixes, respectively) of this list.
 3. Make sure that all exclusion prefixes are contained within existing
    prefixes, unless the list of allowed elements is empty.
 4. Make sure exclusion prefixes in a list don't exclude all of the
    prefixes in the associated prefixes list.
 5. Make sure we have the same number of addresses available on each
    side (public/private), taking exclusion prefixes into account.

Add related ConfigError enum variants, as necessary to cover the
different error cases.

Add unit tests to check validation for a number of VpcExpose objects.

Some other unit tests in the crate need minor adjustments, because they
use incorrect VpcExpose objects with exclusion prefixes not covering
allowed prefixes, or with a mismatch between the number of addresses in
the public and private sets (which we reject because of static NAT).

Signed-off-by: Quentin Monnet <qmo@qmon.net>