feat(vpcpeering): Detect VpcExpose prefixes overlap in VpcManifest
When validating a VpcExpose object, we ensure that the prefixes in each
category (ips, as_range, nots, not_as) do not overlap. However, within a
VpcManifest, we must also ensure that prefixes from different VpcExpose
objects do not collide. In particular, we want to make sure that the
available addresses in the "ips" lists do not collide with "ips" lists
from other VpcExpose (or if the prefixes collide, they should also be
covered by exclusion prefixes). Same applies to the "as" lists.

Introduce a function to check overlap between two sets of addresses (one
set = allowed prefixes + exclusion prefixes). See comments in the code
for details on the implementation. Call this function as part of the
validation process for a VpcManifest, to ensure there is no collision.

Signed-off-by: Quentin Monnet <qmo@qmon.net>