feat(vpcpeering): Introduce validation for VpcPeeringTable
In previous commits, we introduced validation for VpcExpose and
VpcManifest objects; now, we need the same for a VpcPeeringTable.

We had checks for overlap between prefixes of a same VpcExpose, or
between prefixes of different VpcExpose within a VpcManifest object. It
turns out we also need to detect overlap between the prefixes of the
exposes from the manifests for a peering related to a given VPC and the
existing exposes from manifests from all other peerings in the
VpcPeeringTable object, when they relate to the same VPC.

To detect overlap, we reuse the dedicated function introduced in a
previous commit.

On the tests side:

- Add dedicated unit tests, and call the validation function.

- Fix some tests that would incorrectly reuse VpcExpose objects for
  multiple peerings about the same VPC.

Signed-off-by: Quentin Monnet <qmo@qmon.net>