fix(vpcpeering): Do not insert peering with duplicate name
When we try to insert a peering with a name that already exists for
another peering in a VpcPeeringTable, this results in an error, but it
also results in the peering being inserted anyway, as we only detect the
name collision when inserting (and we do not implement rollback).

As a consequence, we keep an incorrect peering entry in the table. This
can cause unexpected collisions on prefixes when trying to add
subsequent, legitimate peerings to the table, for example.

Instead of inserting unconditionally, let's check first for the
existence of the peering name in the table.

Fixes: e45c0ef3ac7a (feat(mgmt): define external model)
Signed-off-by: Quentin Monnet <qmo@qmon.net>