feat(vpcpeering): Introduce validation for VpcManifest
Just like we did for VpcExpose, implement validation for VpcManifest.
There's not much to do this time, because most of the work consists in
validating the VpcExpose objects. Make sure the name is not empty, make
sure the lists of VpcExpose objects are not empty, and done.

Note: We should also prevent prefix overlap between the different
exposes in the manifest; this is not trivial, and will be added as a
separate commit.

Fix some tests that would use manifests with empty VpcExpose lists.

Signed-off-by: Quentin Monnet <qmo@qmon.net>