ci: Add workflow to run a FOSSA scan
Apparently the "Quick Import" feature from FOSSA is not good enough for
Rust projects, and we need to run the FOSSA CLI tool, either ourselves
or using the official GitHub Action. This commit introduces a new
workflow that uses this Action to perform a scan on Pull Requests and on
pushes to the main branch.

Signed-off-by: Quentin Monnet <qmo@qmon.net>