fix: move vpc-manager into mgmt
This always belonged here.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>

fix: validate interface names in mgmt

InterfaceName is a restricted version of String.  By allowing String to be specified instead, we expose the interface manager to all manor of bug paths which we can avoid.  The solution is to statically ensure that all code paths which lead to rtnetlink calls are validated, and this involves a wrapper type.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>

fix: clean up imports

Remove the `#[allow(unused)]` as it no longer obtains

Signed-off-by: Daniel Noland <daniel@githedgehog.com>

fix: demote VpcId new to an unchecked private method

Making this method public is dangerous as it allows the creation of invalid interface names by accident.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>

feat: impl TryFrom<NonZero<u32>> for RouteTableId

This is useful in the next commit.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>

fix: align RouteTableId type

This is part of an ongoing effort to align the types between routing and net crates

Signed-off-by: Daniel Noland <daniel@githedgehog.com>

fix: more InterfaceName type alignment

Signed-off-by: Daniel Noland <daniel@githedgehog.com>

fix: trivial cleanup

Signed-off-by: Daniel Noland <daniel@githedgehog.com>

feat: derive MultiIndexMap for Vpc

This will make it easier to convert a Vpc table into a requirements table.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>

feat: replace VpcTable with MultiIndexMap derivation

The functionality here is redundant.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>

fix: trivial comment cleanup

Signed-off-by: Daniel Noland <daniel@githedgehog.com>

fix: improve type precision in error

Signed-off-by: Daniel Noland <daniel@githedgehog.com>

fix: avoid bookkeeping footgun in api

We can skip the possibility that the idmap is wrong here.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>

Signed-off-by: Daniel Noland <daniel@githedgehog.com>

Signed-off-by: Daniel Noland <daniel@githedgehog.com>

Signed-off-by: Daniel Noland <daniel@githedgehog.com>

feat: add tracing instrumentation

Signed-off-by: Daniel Noland <daniel@githedgehog.com>

WIP

Signed-off-by: Daniel Noland <daniel@githedgehog.com>

WIP

Signed-off-by: Daniel Noland <daniel@githedgehog.com>