fix: validate interface names in mgmt
InterfaceName is a restricted version of String.  By allowing String to be specified instead, we expose the interface manager to all manor of bug paths which we can avoid.  The solution is to statically ensure that all code paths which lead to rtnetlink calls are validated, and this involves a wrapper type.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>