feat(mgmt): empty peering table & move validations
The vpc peering table holds the VPC peerings desired for a certain
overlay configuration. Soon in the processing of a config, the
peerings are validated and moved to each of the VPCs participating.
Therefore, we can empty it since it is no longer needed.

The gRPC server uses the VpcPeering builders to deserialize the
peerings into them. Since this happens before the config is
actually processed (other than building it from gRPC), validation
should occur later, be it for the sake of clarity in logs.

Closes: #128
Closes: #425

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>