feat(nat): Add new PrefixTrie implementation
Add an updated copy of the PrefixTrie implementation that we use for NAT
into the mgmt crate.

Here is how the copies differ:

- The new implementation for PrefixTrie<T> is generic and supports
  values other than Strings.
- The find() and find_ip() methods have been collapsed into a new
  lookup() method that takes an IP address as a parameter and returns
  the matching Prefix (in addition to the value) from the table lookup.
- Tracing macros that would decorate all functions have been removed.
- Some formatting have been improved, in particular the comments have
  been rewrapped with a line width of 100 characters.

The presence of the two copies in the repository is only temporary, we
do not need the two of them. The initial copy has been kept to avoid
breaking compilation in the dataplane crate, it will be removed in a
subsequent commit.

Signed-off-by: Quentin Monnet <qmo@qmon.net>