feat(nat): Add code to insert peering into NAT rules table
Add a function, in a separate submodule, to convert a peering into new
entries in a table with rules for static NAT.

Note: add_peering(table, peering) should be part of PerVniTable, but we
prefer to keep it in a separate submodule because it relies on
definitions from the external models, unlike the rest of the PerVniTable
implementation.

Also add functions to "optimise" a VpcExpose object _in the context of
NAT_, by collapsing allowed and excluded prefixes when possible.

Signed-off-by: Quentin Monnet <qmo@qmon.net>