feat(nat): Provide new IpList implementation
The previous IpList implementation would consider lists of prefixes such
that we would do a 1-to-1 mapping from one prefix into another prefix.

This new version assumes that, insead, we're doing a 1-to-1 mapping
between a list of prefixes and another list of prefixes, while taking
into account exclusion prefixes to apply to these lists.

This new IpList structure will be incorporated into the NAT logics in a
subsequent commit.

Signed-off-by: Quentin Monnet <qmo@qmon.net>