feat(vpcpeering): Introduce validation for VpcTable
In previous commits, we introduced validation for VpcExpose and
VpcManifest objects; now we do the same for a VpcTable.

Here, we simply need to ensure, after building the table, that for any
two given Vpcs, we have at most a single Peering object between the two:
having more than one is not currently supported.

Signed-off-by: Quentin Monnet <qmo@qmon.net>