feat(nat): Optimize peerings before adding them to NAT tables
Inside of the NAT code, leverage the optimisation function to
(potentially) reduce the number of prefixes and exclusion prefixes
inside of the VpcExpose objects of a Peering before we process this
Peering to extend the NAT tables.

Now that we call the optimize_peering() function from add_peering(),
make it private: it's unlikely we'll need to call it from somewhere
else.

Signed-off-by: Quentin Monnet <qmo@qmon.net>