feat(mgmt): strong typing for derived interface names
Linux interface names have strict rules.
These rules are captured and enforced by the `InterfaceName` type.

This commit switches the type of the generated interface names to indicate this to downstream consumers (including the VPC manager).

To be clear, these constraints are absolutely accounted for already in the generated string.  We just need the types to catch up.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>